==> src/00_get_ice_data.R <==
#List of mountain lakes from literature search that included in situ data: 
#
#1. Albion Lake and Silver Lake: 
#From Christianson et al., 2021 - On thin ice: Linking elevation and long-term losses of lake ice cover
#doi: https://doi.org/10.1002/lol2.10181.
#Data are available in the Niwot Ridge LTER repository at: https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-nwt.106.3
#
#2. Castle Lake: 
#From Castle Lake shared dropbox (ice out only)
#
#3. Lake Lunz:
#From Kainz et al., 2017 - Irregular changes in lake surface water temperature and ice cover in subalpine Lake Lunz, Austria
#doi: 10.1080/20442041.2017.1294332
#Data not available through paper/repository. I received the data through email by corresponding with Martin Kainz
#
#4. Lake Morskie Oko:
#From Pociask-Karteczka et al., 2012 - Recent trends ni ice cover duration for Lake Morskie Oko (Tatra Mountains, East-Central Europe)
#doi: 10.2166/nh.2012.019
#Received data from personal communication with Instytut Meteorologii i Gospodarki Wodnej - Panstwowy Instytut Badawczy
#Or - Institute of Meteorology and Water Management
#Contact person Anna Szewczykowska
#
#5. Lej de San Murezzan (St. Moritz, Switzerland):
#From Global Lake and River Ice Phenology Dataset: https://nsidc.org/data/G01377/versions/1
#Benson, B., J. Magnuson, and S. Sharma. 2000, updated 2020. 
#Global Lake and River Ice Phenology Database, Version 1. 
#Freeze/Thaw Dates. Boulder, Colorado USA. NSIDC: National Snow and Ice Data Center. 
#doi: https://doi.org/10.7265/N5W66HP8. 2022.03.23.
#
#6. Norwegian dataset (13 lakes above 1000m)
#From:  L’Abée-Lund, et al., 2021. 
#Geographic variation and temporal trends in ice phenology in Norwegian lakes during the period 1890–2020.
#dpo: https://doi.org/10.5194/tc-15-2333-2021
#Dryad dataset from supplemental material https://datadryad.org/stash/dataset/doi:10.5061/dryad.bk3j9kd9x
